% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{biplot.bfa}
\alias{biplot.bfa}
\title{Display a biplot}
\usage{
\method{biplot}{bfa}(x, factors = c(1, 2), ...)
}
\arguments{
\item{x}{A bfa object}

\item{factors}{Integer vector giving indices of the factors to plot}

\item{...}{Additional arguments to biplot; see \code{?biplot}}
}
\value{
Displays a biplot
}
\description{
Display a biplot
}

