% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_coda.R
\name{get_coda}
\alias{get_coda}
\title{Get coda object}
\usage{
get_coda(model, loadings = TRUE, scores = FALSE, scale = attr(model,
  "type") != "gauss", positive = FALSE)
}
\arguments{
\item{model}{a \code{bfa} model}

\item{loadings}{Return samples of the factor loadings? (default is TRUE)}

\item{scores}{Return samples of the factor scores? (default is FALSE)}

\item{scale}{Return factor loadings on the correlation scale? (default is TRUE for mixed/copula models, FALSE for Gaussian models)}

\item{positive}{Post-process to enforce positivity constraints}
}
\value{
An \code{mcmc} object
}
\description{
Returns an \code{mcmc} object for use in \code{coda} functions for
convergence diagnostics, HPD intervals, etc.
}

