% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postproc.R
\name{get_posterior_loadings}
\alias{get_posterior_loadings}
\title{Get posterior samples of factor loadings}
\usage{
get_posterior_loadings(model, scale = attr(model, "type") != "gauss")
}
\arguments{
\item{model}{a \code{bfa} model}

\item{scale}{Return factor loadings on the correlation scale? (default is TRUE for mixed/copula models, FALSE for Gaussian models)}
}
\value{
A p x k x (no. of mcmc samples) array containing samples of the p x k loadings matrix
}
\description{
Returns an array of posterior samples of factor loadings
}

