% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postproc.R
\name{get_posterior_scores}
\alias{get_posterior_scores}
\title{Get posterior samples of factor scores}
\usage{
get_posterior_scores(model)
}
\arguments{
\item{model}{a \code{bfa} model}
}
\value{
An \code{mcmc} object
}
\description{
Returns an array of posterior samples of factor scores
}

