% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class.R
\name{mean.bfa}
\alias{mean.bfa}
\title{Extract posterior means from bfa object}
\usage{
\method{mean}{bfa}(x, scale = attr(x, "type") != "gauss", ...)
}
\arguments{
\item{x}{A bfa object}

\item{scale}{Return factor loadings on the correlation scale? (default is TRUE for mixed/copula models, FALSE for Gaussian models)}

\item{...}{Ignored}
}
\value{
A list with elements loadings and scores containing MCMC sample means
}
\description{
Extract posterior means from bfa object
}

