\name{var.bfa}
\alias{var.bfa}
\title{Extract posterior means from bfa object}
\usage{
  var.bfa(x, ...)
}
\arguments{
  \item{x}{A bfa object}

  \item{...}{Ignored}
}
\value{
  A list with elements loadings and scores containing MCMC
  sample variances
}
\description{
  Extract posterior means from bfa object
}

