% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postproc.R
\name{woodbury}
\alias{woodbury}
\title{Invert the matrix lam*t(lam)+diag(u) via Woodbury identity}
\usage{
woodbury(lam, u)
}
\arguments{
\item{lam}{p by k matrix}

\item{u}{p dimensional vector}
}
\value{
The matrix inverse
}
\description{
Invert the matrix lam*t(lam)+diag(u) via Woodbury identity
}

