\docType{package}
\name{bfa-package}
\alias{bfa-package}
\title{Bayesian Factor Analysis}
\description{
  Bayesian Factor Analysis
}
\details{
  \tabular{ll}{ Package: \tab bfa\cr Type: \tab Package\cr
  Version: \tab 0.2.3\cr Date: \tab 2012-3-14\cr License:
  \tab GPL-3\cr LazyLoad: \tab yes\cr }

  This package provides model fitting for several Bayesian
  factor models including Gaussian, ordinal probit, mixed
  and semiparametric Gaussian copula factor models under a
  range of priors.
}
\author{
  Jared Murray \email{jared.murray@stat.duke.edu}
}
\keyword{package}

