% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bfa-package.R
\docType{package}
\name{bfa-package}
\alias{bfa-package}
\title{Bayesian Factor Analysis}
\description{
Bayesian Factor Analysis
}
\details{
\tabular{ll}{
Package: \tab bfa\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2016-9-07\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}

This package provides model fitting for several Bayesian factor models including
Gaussian, ordinal probit, mixed and semiparametric Gaussian copula factor models
under a range of priors.
}
\author{
Jared Murray \email{jsmurray@stat.cmu.edu}
}
\keyword{package}

