\name{cov_samp}
\alias{cov_samp}
\title{Compute samples of the covariance matrix}
\usage{
  cov_samp(model)
}
\arguments{
  \item{model}{\code{bfa} model object}
}
\value{
  A p x p x (no. of mcmc samples) array containing samples
  of the covariance matrix
}
\description{
  Compute samples of the covariance matrix
}

