\name{mean.bfa}
\alias{mean.bfa}
\title{Extract posterior means from bfa object}
\usage{
\method{mean}{bfa}(x, ...)
}
\arguments{
  \item{x}{A bfa object}

  \item{...}{Ignored}
}
\value{
A list with elements loadings and scores containing MCMC
sample means
}
\description{
Extract posterior means from bfa object
}

