\name{utri_restrict}
\alias{utri_restrict}
\title{Upper triangular loadings restriction}
\usage{
  utri_restrict(p, k)
}
\arguments{
  \item{p}{number of rows (variables)}

  \item{k}{number of columns (factors)}
}
\value{
  p by k matrix
}
\description{
  Returns the restriction matrix corresponding to Geweke &
  Zhou (1996) upper-triangular positive diagonal condition
  '
}

